+++ docs/js/class_p.js
@@ -157,7 +157,7 @@ class Puzzle {
             ["\"__a\"", "z_"],
             ["null", "zO"],
         ];
-        this.version = [2, 26, 11]; // Also defined in HTML Script Loading in header tag to avoid Browser Cache Problems
+        this.version = [2, 26, 13]; // Also defined in HTML Script Loading in header tag to avoid Browser Cache Problems
         this.undoredo_disable = false;
         this.comp = false;
         this.multisolution = false;
@@ -166,6 +166,14 @@ class Puzzle {
         this.conflicts = new Conflicts();
         this.previous_sol = [];
         this.conflict_cells = [];
+        this.url = [];
+        this.ignored_line_types = {
+            2: 1, // Black color
+            5: 1, // Grey Color
+            80: 1, // Thin
+            12: 1, // Dotted
+            13: 1 // Fat dots
+        };
     }
 
     reset() {
@@ -369,8 +377,9 @@ class Puzzle {
         // pause-unpause layer
         let pause_canvas = document.getElementById("pause_canvas");
         let pause_ctx = pause_canvas.getContext("2d");
-        pause_canvas.style.width = this.canvasx.toString() + "px";
-        pause_canvas.style.height = this.canvasy.toString() + "px";
+        let factor = this.resol * 5;
+        pause_canvas.style.width = (this.canvasx - factor).toString() + "px";
+        pause_canvas.style.height = (this.canvasy - factor).toString() + "px";
         pause_canvas.width = this.resol * this.canvasx;
         pause_canvas.height = this.resol * this.canvasy;
     }
@@ -382,7 +391,7 @@ class Puzzle {
 
         // set the style and font
         pause_ctx.filleStyle = Color.BLUE;
-        let font_size = 0.09 * pause_canvas.height; // 90 % of display size/ height of canvas
+        let font_size = 0.09 * pause_canvas.height; // 9 % of display size/ height of canvas
         pause_ctx.font = font_size + 'px sans-serif';
         let lineheight = 1.2 * font_size;
         let textstring = "Paused\nClick on \"Start\"\nor \"F4\"";
@@ -1984,11 +1993,12 @@ class Puzzle {
 
     point_usecheck() {
         for (var i in this.point) {
-            if (this.point[i].use === -1) {;
-            } else if (this.point[i].x < this.margin || this.point[i].x > this.canvasx - this.margin || this.point[i].y < this.margin || this.point[i].y > this.canvasy - this.margin) {
-                this.point[i].use = 0;
-            } else {
-                this.point[i].use = 1;
+            if (this.point[i].use !== -1) {
+                if (this.point[i].x < this.margin || this.point[i].x > this.canvasx - this.margin || this.point[i].y < this.margin || this.point[i].y > this.canvasy - this.margin) {
+                    this.point[i].use = 0;
+                } else {
+                    this.point[i].use = 1;
+                }
             }
         }
     }
@@ -2087,10 +2097,11 @@ class Puzzle {
             this.redraw(true); // Reflects SVG elements
             this.ctx = old_canvas;
             this.redraw(); // Back to original display
+            this.mode[this.mode.qa].edit_mode = mode; // retain original mode
 
             return svg_canvas.getSerializedSvg(true);
         }
-        this.mode[this.mode.qa].edit_mode = mode;
+        this.mode[this.mode.qa].edit_mode = mode; // retain original mode
 
         if (document.getElementById("nb_margin2").checked) {
             this.canvasx = cx;
@@ -2656,44 +2667,6 @@ class Puzzle {
                     this[this.mode.qa + "_col"][this.mode[this.mode.qa][this.mode[this.mode.qa].edit_mode][0]] = [];
                 }
                 break;
-            case "combi":
-                switch (this.mode[this.mode.qa][this.mode[this.mode.qa].edit_mode][0]) {
-                    case "tents":
-                        break;
-                    case "linex":
-                        break;
-                    case "edgex":
-                        break;
-                    case "edgexoi":
-                        break;
-                    case "blpo":
-                        break;
-                    case "blwh":
-                        break;
-                    case "battleship":
-                        break;
-                    case "star":
-                        break;
-                    case "magnets":
-                        break;
-                    case "lineox":
-                        break;
-                    case "yajilin":
-                        break;
-                    case "hashi":
-                        break;
-                    case "arrowS":
-                        break;
-                    case "shaka":
-                        break;
-                    case "numfl":
-                        break;
-                    case "alfl":
-                        break;
-                    case "edgesub":
-                        break;
-                }
-                break;
         }
         this.redraw();
     }
@@ -2834,11 +2807,7 @@ class Puzzle {
             text = text.split(this.replace[i][0]).join(this.replace[i][1]);
         }
 
-        var u8text = new TextEncoder().encode(text);
-        var deflate = new Zlib.RawDeflate(u8text);
-        var compressed = deflate.compress();
-        var char8 = Array.from(compressed, e => String.fromCharCode(e)).join("");
-        var ba = window.btoa(char8);
+        var ba = encrypt_data(text);
         var url = location.href.split('?')[0];
         // console.log("save",text.length,"=>",compressed.length,"=>",ba.length); //Github ba.length max 7360
 
@@ -2993,11 +2962,7 @@ class Puzzle {
             text = text.split(this.replace[i][0]).join(this.replace[i][1]);
         }
 
-        var u8text = new TextEncoder().encode(text);
-        var deflate = new Zlib.RawDeflate(u8text);
-        var compressed = deflate.compress();
-        var char8 = Array.from(compressed, e => String.fromCharCode(e)).join("");
-        var ba = window.btoa(char8);
+        var ba = encrypt_data(text);
         var url = location.href.split('?')[0];
         let solution_clone;
         // if solution exist then copy the solution as well
@@ -3007,11 +2972,7 @@ class Puzzle {
             } else {
                 solution_clone = this.solution;
             }
-            u8text = new TextEncoder().encode(solution_clone);
-            deflate = new Zlib.RawDeflate(u8text);
-            compressed = deflate.compress();
-            char8 = Array.from(compressed, e => String.fromCharCode(e)).join("");
-            var ba_s = window.btoa(char8);
+            var ba_s = encrypt_data(solution_clone);
             // Warning Long URL
             if ((ba.length + ba_s.length) >= 7360) {
                 Swal.fire({
@@ -3149,11 +3110,7 @@ class Puzzle {
             text = text.split(this.replace[i][0]).join(this.replace[i][1]);
         }
 
-        var u8text = new TextEncoder().encode(text);
-        var deflate = new Zlib.RawDeflate(u8text);
-        var compressed = deflate.compress();
-        var char8 = Array.from(compressed, e => String.fromCharCode(e)).join("");
-        var ba = window.btoa(char8);
+        var ba = encrypt_data(text);
         var url = location.href.split('?')[0];
         //console.log("save",text.length,"=>",compressed.length,"=>",ba.length);
 
@@ -3275,11 +3232,7 @@ class Puzzle {
             text = text.split(this.replace[i][0]).join(this.replace[i][1]);
         }
 
-        var u8text = new TextEncoder().encode(text);
-        var deflate = new Zlib.RawDeflate(u8text);
-        var compressed = deflate.compress();
-        var char8 = Array.from(compressed, e => String.fromCharCode(e)).join("");
-        var ba = window.btoa(char8);
+        var ba = encrypt_data(text);
         var url = location.href.split('?')[0];
         //console.log("save",text.length,"=>",compressed.length,"=>",ba.length);
 
@@ -3300,11 +3253,7 @@ class Puzzle {
         var text;
         text = JSON.stringify(this.make_solution());
 
-        var u8text = new TextEncoder().encode(text);
-        var deflate = new Zlib.RawDeflate(u8text);
-        var compressed = deflate.compress();
-        var char8 = Array.from(compressed, e => String.fromCharCode(e)).join("");
-        var ba = window.btoa(char8);
+        var ba = encrypt_data(text);
         //console.log("save",text.length,"=>",compressed.length,"=>",ba.length);
 
         // Warning Long URL
@@ -3394,13 +3343,7 @@ class Puzzle {
             if (document.getElementById("sol_loopline").checked === true || checkall) {
                 if (document.getElementById("sol_ignoreloopline").checked === true) {
                     for (var i in this[pu].line) {
-                        if ((this["pu_q"].line[i] &&
-                                (this["pu_q"].line[i] === 2 || // Black color
-                                    this["pu_q"].line[i] === 5 || // Grey Color
-                                    this["pu_q"].line[i] === 80 || // Thin
-                                    this["pu_q"].line[i] === 12 || // Dotted
-                                    this["pu_q"].line[i] === 13 // Fat dots
-                                ))) {
+                        if (this["pu_q"].line[i] && this.ignored_line_types[this["pu_q"].line[i]]) {
                             // Ignore the line
                         } else {
                             if (this[pu].line[i] === 3) {
@@ -3422,13 +3365,7 @@ class Puzzle {
 
                 if (document.getElementById("sol_ignoreloopline").checked === true) {
                     for (var i in this[pu].freeline) {
-                        if ((this["pu_q"].freeline[i] &&
-                                (this["pu_q"].freeline[i] === 2 || // Black color
-                                    this["pu_q"].freeline[i] === 5 || // Grey Color
-                                    this["pu_q"].freeline[i] === 80 || // Thin
-                                    this["pu_q"].freeline[i] === 12 || // Dotted
-                                    this["pu_q"].freeline[i] === 13 // Fat dots
-                                ))) {
+                        if (this["pu_q"].freeline[i] && this.ignored_line_types[this["pu_q"].freeline[i]]) {
                             // Ignore the line
                         } else {
                             if (this[pu].freeline[i] === 3) {
@@ -3693,13 +3630,7 @@ class Puzzle {
                             break;
                         case "loopline":
                             for (var i in this[pu].line) {
-                                if ((this["pu_q"].line[i] &&
-                                        (this["pu_q"].line[i] === 2 || // Black color
-                                            this["pu_q"].line[i] === 5 || // Grey Color
-                                            this["pu_q"].line[i] === 80 || // Thin
-                                            this["pu_q"].line[i] === 12 || // Dotted
-                                            this["pu_q"].line[i] === 13 // Fat dots
-                                        ))) {
+                                if (this["pu_q"].line[i] && this.ignored_line_types[this["pu_q"].line[i]]) {
                                     // Ignore the line
                                 } else {
                                     if (this[pu].line[i] === 3) {
@@ -3711,13 +3642,7 @@ class Puzzle {
                             }
 
                             for (var i in this[pu].freeline) {
-                                if ((this["pu_q"].freeline[i] &&
-                                        (this["pu_q"].freeline[i] === 2 || // Black color
-                                            this["pu_q"].freeline[i] === 5 || // Grey Color
-                                            this["pu_q"].freeline[i] === 80 || // Thin
-                                            this["pu_q"].freeline[i] === 12 || // Dotted
-                                            this["pu_q"].freeline[i] === 13 // Fat dots
-                                        ))) {
+                                if (this["pu_q"].freeline[i] && this.ignored_line_types[this["pu_q"].freeline[i]]) {
                                     // Ignore the line
                                 } else {
                                     if (this[pu].freeline[i] === 3) {
@@ -5106,6 +5031,24 @@ class Puzzle {
                                 }
                             }
                             text += "\n";
+                        }
+                    }
+                }
+            } else if (header === "tapa_contest" || header === "tc") {
+                // Answer - Shading
+                if (!isEmpty(this.pu_a.surface)) {
+                    for (var j = 2; j < this.ny0 - 2; j++) {
+                        for (var i = 2; i < this.nx0 - 2; i++) {
+                            // any shades of grey including black
+                            if (this.pu_a.surface[i + j * (this.nx0)] &&
+                                (this.pu_a.surface[i + j * (this.nx0)] === 1 ||
+                                    this.pu_a.surface[i + j * (this.nx0)] === 8 ||
+                                    this.pu_a.surface[i + j * (this.nx0)] === 3 ||
+                                    this.pu_a.surface[i + j * (this.nx0)] === 4)) {
+                                text += "1";
+                            } else {
+                                text += "0";
+                            }
                         }
                     }
                 }
@@ -9889,14 +9832,41 @@ class Puzzle {
                     this.drawing_mode = 50;
                 }
             } else {
-                if (!this[this.mode.qa].lineE[num]) { // Insert cross
-                    this.record(symboltype, num);
-                    this[this.mode.qa].lineE[num] = 98;
-                    this.drawing_mode = 52;
-                } else if (this[this.mode.qa].lineE[num] === 98) { // Remove Cross
-                    this.record(symboltype, num);
-                    delete this[this.mode.qa].lineE[num];
-                    this.drawing_mode = 50;
+                // Ignore if edge already exist
+                // Do this only for square grids for now
+                if (this.gridtype === "square") {
+
+                    let neighbor1 = this.point[num].neighbor[0];
+                    let neighbor2 = this.point[num].neighbor[1];
+                    let edge_num;
+                    let corners = this.point[neighbor2].surround;
+
+                    // If difference is 1 then its left and right else its top and bottom
+                    if (Math.abs(neighbor1 - neighbor2) === 1) {
+                        edge_num = corners[0].toString() + "," + corners[3].toString();
+                    } else {
+                        edge_num = corners[0].toString() + "," + corners[1].toString();
+                    }
+
+                    if (!this[this.mode.qa].lineE[num] && !this[this.mode.qa].lineE[edge_num]) { // Insert cross
+                        this.record(symboltype, num);
+                        this[this.mode.qa].lineE[num] = 98;
+                        this.drawing_mode = 52;
+                    } else if (this[this.mode.qa].lineE[num] === 98) { // Remove Cross
+                        this.record(symboltype, num);
+                        delete this[this.mode.qa].lineE[num];
+                        this.drawing_mode = 50;
+                    }
+                } else {
+                    if (!this[this.mode.qa].lineE[num]) { // Insert cross
+                        this.record(symboltype, num);
+                        this[this.mode.qa].lineE[num] = 98;
+                        this.drawing_mode = 52;
+                    } else if (this[this.mode.qa].lineE[num] === 98) { // Remove Cross
+                        this.record(symboltype, num);
+                        delete this[this.mode.qa].lineE[num];
+                        this.drawing_mode = 50;
+                    }
                 }
             }
         } else {
@@ -9968,15 +9938,43 @@ class Puzzle {
             this.last = num;
             this.redraw();
         } else if ((this.point[num].type === 2 || this.point[num].type === 3)) {
-            if (this.drawing_mode == 52) {
-                if (!this[this.mode.qa].lineE[num]) { // Insert cross
-                    this.record("lineE", num);
-                    this[this.mode.qa].lineE[num] = 98;
+            if (this.gridtype === "square") {
+                // Ignore if edge already exist
+                // Do this only for square grids for now
+                let neighbor1 = this.point[num].neighbor[0];
+                let neighbor2 = this.point[num].neighbor[1];
+                let edge_num;
+                let corners = this.point[neighbor2].surround;
+
+                // If difference is 1 then its left and right else its top and bottom
+                if (Math.abs(neighbor1 - neighbor2) === 1) {
+                    edge_num = corners[0].toString() + "," + corners[3].toString();
+                } else {
+                    edge_num = corners[0].toString() + "," + corners[1].toString();
                 }
-            } else if (this.drawing_mode == 50) {
-                if (this[this.mode.qa].lineE[num] === 98) { // Remove Cross
-                    this.record("lineE", num);
-                    delete this[this.mode.qa].lineE[num];
+
+                if (this.drawing_mode == 52) {
+                    if (!this[this.mode.qa].lineE[num] && !this[this.mode.qa].lineE[edge_num]) { // Insert cross
+                        this.record("lineE", num);
+                        this[this.mode.qa].lineE[num] = 98;
+                    }
+                } else if (this.drawing_mode == 50) {
+                    if (this[this.mode.qa].lineE[num] === 98) { // Remove Cross
+                        this.record("lineE", num);
+                        delete this[this.mode.qa].lineE[num];
+                    }
+                }
+            } else {
+                if (this.drawing_mode == 52) {
+                    if (!this[this.mode.qa].lineE[num]) { // Insert cross
+                        this.record("lineE", num);
+                        this[this.mode.qa].lineE[num] = 98;
+                    }
+                } else if (this.drawing_mode == 50) {
+                    if (this[this.mode.qa].lineE[num] === 98) { // Remove Cross
+                        this.record("lineE", num);
+                        delete this[this.mode.qa].lineE[num];
+                    }
                 }
             }
             this.redraw();
@@ -11511,36 +11509,31 @@ class Puzzle {
         for (var j = r_start; j < (size + r_start); j++) { //  row
             for (var i = c_start; i < (size + c_start); i++) { // column
 
+                let primary = this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)];
+                let secondary = this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)];
+                let checklist = {};
+
                 if (document.getElementById("ignore_pencilmarks").checked) {
-                    var ifcondition = [this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)] &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "2") &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "4") &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "5") &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "6") &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "10"),
-                        this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)] &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "2") &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "4") &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "5") &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "6") &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "10")
-                    ];
+                    checklist = {
+                        2: 1,
+                        4: 1,
+                        5: 1,
+                        6: 1,
+                        10: 1
+                    };
                 } else {
-                    var ifcondition = [this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)] &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "2") &&
-                        (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "4"),
-                        this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)] &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "2") &&
-                        (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] !== "4")
-                    ];
+                    checklist = {
+                        2: 1,
+                        4: 1
+                    };
                 }
 
-                if (ifcondition[0]) {
-                    if (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][2] === "7") {
+                if (primary && !checklist[primary[2]]) {
+                    if (primary[2] === "7") {
                         var sum = 0,
                             a;
                         for (var k = 0; k < 10; k++) {
-                            if (this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][0][k] === 1) {
+                            if (primary[0][k] === 1) {
                                 sum += 1;
                                 a = k + 1;
                             }
@@ -11551,18 +11544,18 @@ class Puzzle {
                             outputstring += '0';
                         }
                     } else {
-                        if (isNaN(parseInt(this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][0]))) {
+                        if (isNaN(parseInt(primary[0]))) {
                             outputstring += '0';
                         } else {
-                            outputstring += this[mode_order[0]].number[(i + 2) + ((j + 2) * this.nx0)][0];
+                            outputstring += primary[0];
                         }
                     }
-                } else if (ifcondition[1]) {
-                    if (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][2] === "7") {
+                } else if (secondary && !checklist[secondary[2]]) {
+                    if (secondary[2] === "7") {
                         var sum = 0,
                             a;
                         for (var k = 0; k < (size + 1); k++) {
-                            if (this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][0][k] === 1) {
+                            if (secondary[0][k] === 1) {
                                 sum += 1;
                                 a = k + 1;
                             }
@@ -11573,10 +11566,10 @@ class Puzzle {
                             outputstring += '0';
                         }
                     } else {
-                        if (isNaN(parseInt(this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][0]))) {
+                        if (isNaN(parseInt(secondary[0]))) {
                             outputstring += '0';
                         } else {
-                            outputstring += this[mode_order[1]].number[(i + 2) + ((j + 2) * this.nx0)][0];
+                            outputstring += secondary[0];
                         }
                     }
                 } else {
@@ -11647,19 +11640,19 @@ class Puzzle {
             document.getElementById(i).style.display = (displaytype === 'inline-block') ? 'table-row' : displaytype;
         }
         for (var i of penpa_modes["square"]['ms']) {
-            document.getElementById("ms_" + i).style.display = displaytype;
+            document.getElementById("ms_" + i).parentElement.style.display = (displaytype === 'inline-block') ? 'list-item' : displaytype;
         }
         for (var i of penpa_modes["square"]['ms1']) {
-            document.getElementById("ms1_" + i).style.display = displaytype;
+            document.getElementById("ms1_" + i).parentElement.style.display = (displaytype === 'inline-block') ? 'list-item' : displaytype;
         }
         for (var i of penpa_modes["square"]['ms3']) {
-            document.getElementById("ms3_" + i).style.display = displaytype;
+            document.getElementById("ms3_" + i).parentElement.style.display = (displaytype === 'inline-block') ? 'list-item' : displaytype;
         }
         for (var i of penpa_modes["square"]['shapemodes']) {
             document.getElementById(i).style.display = displaytype;
         }
         for (var i of penpa_modes["square"]['combisub']) {
-            document.getElementById("combisub_" + i).style.display = displaytype;
+            document.getElementById("combisub_" + i).parentElement.style.display = (displaytype === 'inline-block') ? 'list-item' : displaytype;
         }
         for (var i of penpa_modes["square"]['subcombi']) {
             document.getElementById(i).style.display = displaytype;